pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "semmy"
description = "A tool to measure pictures with scale_bars"
requires-python = ">=3.11"
license = {file = "LICENSE"}
dependencies = [
    'numpy',
    'vispy',
    'PyQt6',
    'pyopengl',
    'pyyaml',
    'opencv-python',
    'importlib-metadata; python_version<"3.11"',
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Utilities",
]
dynamic = ["version", "readme"]

[tool.setuptools.dynamic]
version = {attr = "semmy.__version__"}
readme = {file = "README.md"}

[tool.setuptools.packages.find]
where = ["src"]